requirements.txt
Package                                  Version
---------------------------------------- -----------
aiohappyeyeballs                         2.6.1
aiohttp                                  3.13.2
aiosignal                                1.4.0
annotated-types                          0.7.0
anyio                                    4.12.0
async-timeout                            4.0.3
attrs                                    25.4.0
backoff                                  2.2.1
bcrypt                                   5.0.0
build                                    1.3.0
cachetools                               6.2.4
certifi                                  2025.11.12
cffi                                     2.0.0
charset-normalizer                       3.4.4
chromadb                                 1.4.0
click                                    8.1.8
coloredlogs                              15.0.1
cryptography                             46.0.3
dataclasses-json                         0.6.7
distro                                   1.9.0
durationpy                               0.10
exceptiongroup                           1.3.1
faiss-cpu                                1.13.0
filelock                                 3.19.1
flatbuffers                              25.12.19
frozenlist                               1.8.0
fsspec                                   2025.10.0
google-auth                              2.45.0
googleapis-common-protos                 1.72.0
grpcio                                   1.76.0
h11                                      0.16.0
hf-xet                                   1.2.0
httpcore                                 1.0.9
httptools                                0.7.1
httpx                                    0.28.1
httpx-sse                                0.4.3
huggingface-hub                          0.36.0
humanfriendly                            10.0
idna                                     3.11
importlib_metadata                       8.7.1
importlib_resources                      6.5.2
Jinja2                                   3.1.6
jiter                                    0.12.0
joblib                                   1.5.3
jsonpatch                                1.33
jsonpointer                              3.0.0
jsonschema                               4.25.1
jsonschema-specifications                2025.9.1
kubernetes                               34.1.0
langchain                                0.3.27
langchain-community                      0.3.31
langchain-core                           0.3.81
langchain-openai                         0.3.35
langchain-text-splitters                 0.3.11
langsmith                                0.4.37
markdown-it-py                           3.0.0
MarkupSafe                               3.0.3
marshmallow                              3.26.2
mdurl                                    0.1.2
mmh3                                     5.2.0
mpmath                                   1.3.0
multidict                                6.7.0
mypy_extensions                          1.1.0
networkx                                 3.2.1
numpy                                    2.0.2
oauthlib                                 3.3.1
onnxruntime                              1.19.2
openai                                   2.14.0
opentelemetry-api                        1.39.1
opentelemetry-exporter-otlp-proto-common 1.39.1
opentelemetry-exporter-otlp-proto-grpc   1.39.1
opentelemetry-proto                      1.39.1
opentelemetry-sdk                        1.39.1
opentelemetry-semantic-conventions       0.60b1
orjson                                   3.11.5
overrides                                7.7.0
packaging                                25.0
pdfminer.six                             20251107
pdfplumber                               0.11.8
pillow                                   11.3.0
pinecone-client                          6.0.0
pinecone-plugin-interface                0.0.7
pip                                      25.3
posthog                                  5.4.0
propcache                                0.4.1
protobuf                                 6.33.2
pyasn1                                   0.6.1
pyasn1_modules                           0.4.2
pybase64                                 1.4.3
pycparser                                2.23
pydantic                                 2.12.5
pydantic_core                            2.41.5
pydantic-settings                        2.11.0
Pygments                                 2.19.2
pypdf                                    6.5.0
pypdfium2                                5.2.0
PyPika                                   0.48.9
pyproject_hooks                          1.2.0
python-dateutil                          2.9.0.post0
python-dotenv                            1.2.1
PyYAML                                   6.0.3
referencing                              0.36.2
regex                                    2025.11.3
requests                                 2.32.5
requests-oauthlib                        2.0.0
requests-toolbelt                        1.0.0
rich                                     14.2.0
rpds-py                                  0.27.1
rsa                                      4.9.1
safetensors                              0.7.0
scikit-learn                             1.6.1
scipy                                    1.13.1
sentence-transformers                    5.1.2
setuptools                               58.0.4
shellingham                              1.5.4
six                                      1.17.0
sniffio                                  1.3.1
SQLAlchemy                               2.0.45
sympy                                    1.14.0
tenacity                                 9.1.2
threadpoolctl                            3.6.0
tiktoken                                 0.12.0
tokenizers                               0.22.1
tomli                                    2.3.0
torch                                    2.8.0
tqdm                                     4.67.1
transformers                             4.57.3
typer                                    0.21.0
typer-slim                               0.21.0
typing_extensions                        4.15.0
typing-inspect                           0.9.0
typing-inspection                        0.4.2
urllib3                                  2.3.0
uvicorn                                  0.39.0
uvloop                                   0.22.1
watchfiles                               1.1.1
websocket-client                         1.9.0
websockets                               15.0.1
yarl                                     1.22.0
zipp                                     3.23.0
zstandard                                0.25.0
